feat(app): show session resource usage
