+++ renku_notebooks/__init__.py
@@ -24,17 +24,16 @@ from apispec import APISpec
 import os
 
 from . import config
-from .api.schemas import (
-    LaunchNotebookRequest,
-    LaunchNotebookResponse,
+from .api.schemas.servers_post import LaunchNotebookRequest
+from .api.schemas.servers_get import (
+    NotebookResponse,
     ServersGetRequest,
     ServersGetResponse,
-    ServerLogs,
-    ServerOptionsUI,
-    FailedParsing,
-    AutosavesList,
-    VersionResponse,
 )
+from .api.schemas.logs import ServerLogs
+from .api.schemas.config_server_options import ServerOptionsChoices
+from .api.schemas.autosave import AutosavesList
+from .api.schemas.version import VersionResponse
 from .api.notebooks import (
     user_servers,
     user_server,
@@ -91,9 +90,11 @@ def create_app():
 
     app.config.from_object(config)
 
-    from .api import blueprints
+    from .api.health import bp as health_bp
+    from .api.notebooks import bp as notebooks_bp
+    from .api.auth import bp as auth_bp
 
-    for bp in blueprints:
+    for bp in [auth_bp, health_bp, notebooks_bp]:
         app.register_blueprint(bp)
 
     # Return validation errors as JSON
@@ -142,12 +143,11 @@ def register_swagger(app):
     )
     # Register schemas
     spec.components.schema("LaunchNotebookRequest", schema=LaunchNotebookRequest)
-    spec.components.schema("LaunchNotebookResponse", schema=LaunchNotebookResponse)
+    spec.components.schema("NotebookResponse", schema=NotebookResponse)
     spec.components.schema("ServersGetRequest", schema=ServersGetRequest)
     spec.components.schema("ServersGetResponse", schema=ServersGetResponse)
     spec.components.schema("ServerLogs", schema=ServerLogs)
-    spec.components.schema("ServerOptionsUI", schema=ServerOptionsUI)
-    spec.components.schema("FailedParsing", schema=FailedParsing)
+    spec.components.schema("ServerOptionsChoices", schema=ServerOptionsChoices)
     spec.components.schema("AutosavesList", schema=AutosavesList)
     spec.components.schema("VersionResponse", schema=VersionResponse)
     # Register endpoints
