+++ renku_notebooks/api/health.py
@@ -1,0 +1,11 @@
+from flask import Blueprint, Response
+
+from .. import config
+
+bp = Blueprint("health_blueprint", __name__)
+
+
+@bp.route("/health")
+def health():
+    """Just a health check path."""
+    return Response("service running under {}".format(config.SERVICE_PREFIX))
