+++ renku_notebooks/api/schemas/version.py
@@ -1,0 +1,24 @@
+from marshmallow import fields, Schema
+
+
+class NotebooksServiceInfo(Schema):
+    """Various notebooks service info."""
+
+    anonymousSessionsEnabled = fields.Boolean(required=True)
+    cloudstorageEnabled = fields.Dict(
+        required=True, keys=fields.String, values=fields.Boolean
+    )
+
+
+class NotebooksServiceVersions(Schema):
+    """Notebooks service version and info."""
+
+    data = fields.Nested(NotebooksServiceInfo, required=True)
+    version = fields.String(required=True)
+
+
+class VersionResponse(Schema):
+    """The response for /version endpoint."""
+
+    name = fields.String(required=True)
+    versions = fields.List(fields.Nested(NotebooksServiceVersions), required=True)
