+++ tests/unit/test_schema_validation.py
@@ -1,7 +1,7 @@
 from marshmallow import ValidationError
 import pytest
 
-from renku_notebooks.api.schemas import UserPodAnnotations
+from renku_notebooks.api.schemas.servers_get import UserPodAnnotations
 
 RENKU_ANNOTATION_PREFIX = "renku.io/"
 JUPYTER_ANNOTATION_PREFIX = "jupyter.org/"
@@ -33,25 +33,3 @@ def test_missing_required_annotation_fails():
     response.pop(f"{RENKU_ANNOTATION_PREFIX}projectName")
     with pytest.raises(ValidationError):
         schema.load(response)
-
-
-def test_value_range_check():
-    from renku_notebooks.api.schemas import _in_range
-
-    # test byte size comparison
-    value_range = {"type": "bytes", "min": "1G", "max": "10G"}
-    # too small
-    assert not _in_range("0.5M", value_range)
-    # too big
-    assert not _in_range("200G", value_range)
-    # just right
-    assert _in_range("5G", value_range)
-
-    # test int comparison
-    value_range = {"type": "int", "min": 1, "max": 10}
-    # too small
-    assert not _in_range(0, value_range)
-    # too big
-    assert not _in_range(100, value_range)
-    # just right
-    assert _in_range(5, value_range)
